% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binge.R
\name{binge}
\alias{binge}
\title{report aggregated summary on binged diary entries (for otr or mhs)}
\usage{
binge(path, study, visit, remove)
}
\description{
report aggregated summary on binged diary entries (for otr or mhs)
}
\examples{
binge("/Users/phoebelam/Desktop/clean", "MHS", 1, 2023)

}
