% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoflex.R
\name{cytoflex}
\alias{cytoflex}
\title{consolidate cytoflex data}
\usage{
cytoflex(path, type = c("emp", "mono"))
}
\description{
consolidate cytoflex data
}
\examples{
cytoflex("/Users/phoebelam/Desktop/cytoflex", type='emp')

}
