% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dupbinge.R
\name{dupbinge}
\alias{dupbinge}
\title{report entries with duplicated entries}
\usage{
dupbinge(path, study, visit, exclude)
}
\description{
report entries with duplicated entries
}
\examples{
dup("/Users/phoebelam/Desktop/clean", "MHS")

}
