% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipr.R
\name{iprcalc.aspe}
\alias{iprcalc.aspe}
\title{ipr calculator based on aspe}
\usage{
iprcalc.aspe(year, familysize, income, result = c("ipr", "threshold"))
}
\description{
return ipr based user input of year, family size, and income
}
