% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luminex.R
\name{luminex}
\alias{luminex}
\title{aggregate luminex output, compute mfi cv's,}
\usage{
luminex(path)
}
\description{
aggregate luminex output, compute mfi cv's,
}
\examples{
luminex("/Users/phoebelam/Desktop/luminex")

}
