% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luminex.v2.R
\name{luminex.v2}
\alias{luminex.v2}
\title{consolidate luminex output for mwmh v2 only}
\usage{
luminex.v2(path)
}
\arguments{
\item{path}{the folder to which all the individual files are saved}
}
\description{
grab mfi, extrapolated values, and cv from individual outputs
}
\examples{
luminex.v2 (path= "/Users/phoebelam/Box/FHRC (Weinberg Box Admin)/NIH R01 My World My Heart Study (MWMH)/Wetlab/Immunoassays/Luminex/Raw Data/MWMH V2 Culture Sup Raw Data")


}
