% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spah.R
\name{luminex_spah}
\alias{luminex_spah}
\title{aggregate luminex output, compute mfi cv's,}
\usage{
luminex_spah(path)
}
\description{
aggregate luminex output, compute mfi cv's,
}
\examples{
luminex_spah("/Users/phoebelam/Desktop/spah")

}
