% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhs.importclean.R
\name{mhs.importclean}
\alias{mhs.importclean}
\title{mhs-specific script for cleaning a ripple v1 export ready for v2 import}
\usage{
mhs.importclean(path, importid, new)
}
\arguments{
\item{path}{in quotes, enter the path for the exported file (make sure this is .csv)}

\item{importid}{in quotes, ripple -> site admin -> import -> under "download template" select the new study v2 study -> select "template" -> scroll to all the way to the right, you can find the import id under the column "importType"}

\item{new}{in quotes, enter where you want and the name of the new cleaned file the function outputs (make sure this is .xlsx)}
}
\description{
takes a ripple-exported v1 .csv file and turn it into a cleaned .xlsx import
}
\examples{
mhs.importclean (path = "/Users/phoebelam/Desktop/ripple v1 to v2 migrate/c2021_v1_export.csv", 
new = "/Users/phoebelam/Desktop/ripple v1 to v2 migrate/c2021_v2_import.xlsx", 
importid = "aDTL2wkzBmMEXsL4u")
 
}
