% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhs.luminex.R
\name{mhs.luminex}
\alias{mhs.luminex}
\title{consolidate luminex output for mhs only}
\usage{
mhs.luminex(path)
}
\arguments{
\item{path}{the folder to which all the individual files are saved}
}
\description{
consolidate mfi and cv from individual outputs
}
\examples{
mhs.luminex (path= "/Users/phoebelam/Box/FHRC/Mentoring & Health Study (MHS)/Wet Lab/Immunoassays/Luminex/Cohort 1 Data")


}
