% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwmh.fitbit.R
\name{mwmh.fitbit}
\alias{mwmh.fitbit}
\title{consolidate individual fitbit data}
\usage{
mwmh.fitbit(path)
}
\arguments{
\item{path}{the folder that contains all the individual files to be merged (data can be in subfolders)}
}
\description{
consolidate individual fitbit data and export activity and sleep data in long and wide format
relevant for mwmh, but can be adapted for studies that use fitbit
}
\examples{
mwmh.fitbit (path = "/Users/phoebelam/Box/FHRC (Weinberg Box Admin)/NIH R01 My World My Heart Study (MWMH)/Data/V2/OG & IP Data by Source/Fitbit Data/MWMH Fitbit_Individual Excel Files")

}
