% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stoplight.R
\name{stoplight}
\alias{stoplight}
\title{aggregate .csv stoplight task files,}
\usage{
stoplight(path)
}
\description{
aggregate .csv stoplight task files,
}
\examples{
stoplight("/Users/phoebelam/Desktop/stoplight")

}
