% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templeella.R
\name{tempella}
\alias{tempella}
\title{consolidate ella output for washu}
\usage{
tempella(path)
}
\arguments{
\item{path}{the folder to which all the individual files are saved}
}
\description{
consolidate ella output for washu
}
