% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/washuella.R
\name{washuella}
\alias{washuella}
\title{consolidate ella output for washu}
\usage{
washuella(path)
}
\arguments{
\item{path}{the folder to which all the individual files are saved}
}
\description{
consolidate ella output for washu
}
