% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luminex.v1.R
\name{luminex}
\alias{luminex}
\title{consolidate luminex output}
\usage{
luminex(path)
}
\arguments{
\item{path}{the folder to which all the individual files are saved}
}
\description{
grab mfi, extrapolated values, and cv from individual outputs
}
\examples{
luminex (path= "/Users/phoebelam/Box/FHRC (Weinberg Box Admin)/NIH R01 My World My Heart Study (MWMH)/Wetlab/Immunoassays/Luminex/Raw Data/MWMH V1 Culture Sup Raw Data")

}
